# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workspaces/{id}": {
      "get": {
        "operationId": "WorkspaceController_findOne",
        "summary": "Get workspace by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "WorkspaceController_update",
        "summary": "Update workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "WorkspaceController_delete",
        "summary": "Delete workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/workspace-templates/{id}": {
      "get": {
        "operationId": "WorkspaceTemplateController_findOne",
        "summary": "Get template details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workspace Templates"
        ]
      }
    },
...
    "/admin/workspaces": {
      "get": {
        "operationId": "AdminWorkspaceController_findAll",
        "summary": "Get all workspaces (Admin only)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "AdminWorkspaceController_create",
        "summary": "Create a new workspace/template as admin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/workspaces/{id}": {
      "patch": {
        "operationId": "AdminWorkspaceController_update",
        "summary": "Update any workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminWorkspaceController_delete",
        "summary": "Delete any workspace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkspaceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspaces"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/workspace-templates": {
      "get": {
        "operationId": "AdminWorkspaceTemplateController_findAll",
        "summary": "List all workspace templates",
        "parameters": [
          {
            "name": "organization_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sport_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "country_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceTemplatesListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspace Templates"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "AdminWorkspaceTemplateController_create",
        "summary": "Create a new workspace template",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceTemplateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspace Templates"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/workspace-templates/{id}": {
      "patch": {
        "operationId": "AdminWorkspaceTemplateController_update",
        "summary": "Update workspace template",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceTemplateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspace Templates"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminWorkspaceTemplateController_delete",
        "summary": "Delete workspace template",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkspaceTemplateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Workspace Templates"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Elbruso API",
    "description": "BI System REST API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Workspace name",
            "example": "My Workspace",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Workspace description",
            "example": "Team collaboration workspace",
            "maxLength": 500
          },
          "icon": {
            "type": "string",
            "description": "Workspace icon (emoji or URL)",
            "example": "🚀"
          },
          "is_template": {
            "type": "boolean",
            "description": "Is this workspace a template",
            "example": false
          }
        },
        "required": [
          "name"
        ]
      },
...
      "WorkspaceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "is_template": {
            "type": "boolean"
          },
          "owner_id": {
            "type": "string"
          },
          "owner_email": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "nullable": true
          },
          "organization_id": {
            "type": "number",
            "nullable": true
          },
          "sport_id": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "is_template",
          "owner_id",
          "owner_email",
          "created_at",
          "updated_at",
          "is_active",
          "metadata",
          "organization_id",
          "sport_id"
        ]
      },
      "WorkspacesListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponseDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Workspace name",
            "example": "Updated Workspace",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Workspace description",
            "maxLength": 500
          },
          "icon": {
            "type": "string",
            "description": "Workspace icon"
          },
          "is_template": {
            "type": "boolean",
            "description": "Is this workspace a template"
          }
        }
      },
...
      "TemplatePageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplatePageResponseDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "order"
        ]
      },
      "TemplateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "previewImageUrl": {
            "type": "string"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplatePageResponseDto"
            }
          },
          "organizationId": {
            "type": "number"
          },
          "sportId": {
            "type": "number"
          },
          "countryId": {
            "type": "number"
          },
          "isDefault": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "category",
          "previewImageUrl",
          "pages",
          "organizationId",
          "sportId",
          "countryId",
          "isDefault",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "DeleteWorkspaceResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "WorkspaceTemplateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "is_template": {
            "type": "boolean"
          },
          "owner_id": {
            "type": "string"
          },
          "owner_email": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "nullable": true
          },
          "organization_id": {
            "type": "number",
            "nullable": true
          },
          "sport_id": {
            "type": "number",
            "nullable": true
          },
          "country_id": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "is_template",
          "owner_id",
          "owner_email",
          "created_at",
          "updated_at",
          "is_active",
          "metadata",
          "organization_id",
          "sport_id",
          "country_id"
        ]
      },
      "WorkspaceTemplatesListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceTemplateResponseDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateWorkspaceTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Template name",
            "example": "Regional Organization Template",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Template description",
            "example": "Standard workspace for regional organizations",
            "maxLength": 500
          },
          "icon": {
            "type": "string",
            "description": "Template icon (emoji or URL)",
            "example": "🏢"
          },
          "organization_id": {
            "type": "number",
            "description": "Organization ID linking"
          },
          "sport_id": {
            "type": "number",
            "description": "Sport ID linking"
          },
          "country_id": {
            "type": "number",
            "description": "Country ID linking (stored in metadata)"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateWorkspaceTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Template name",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Template description",
            "maxLength": 500
          },
          "icon": {
            "type": "string",
            "description": "Template icon"
          },
          "organization_id": {
            "type": "number",
            "description": "Organization ID linking"
          },
          "sport_id": {
            "type": "number",
            "description": "Sport ID linking"
          },
          "country_id": {
            "type": "number",
            "description": "Country ID linking (stored in metadata)"
          }
        }
      },
      "DeleteWorkspaceTemplateResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}